requirements.txt
asgiref==3.7.2
astroid==2.15.8
autopep8==2.0.4
colorama==0.4.6
dill==0.3.7
Django==4.2.5
djangorestframework==3.14.0
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
mysqlclient==2.2.0
Pillow==10.0.1
platformdirs==3.10.0
pycodestyle==2.11.0
pylint==2.17.6
pytz==2023.3.post1
sqlparse==0.4.4
tomlkit==0.12.1
tzdata==2023.3
wrapt==1.15.0
